+++ minishell/libft/ft_calloc.c
@@ -6,11 +6,12 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 18:08:46 by seunan            #+#    #+#             */
-/*   Updated: 2023/03/20 17:36:41 by seunan           ###   ########seoul.kr  */
+/*   Updated: 2023/09/04 22:41:37 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdio.h>
 
 void	*ft_calloc(size_t count, size_t size)
 {
@@ -18,7 +19,10 @@ void	*ft_calloc(size_t count, size_t size)
 
 	mem = malloc(size * count);
 	if (!mem)
-		return (0);
+	{
+		perror("malloc");
+		exit(EXIT_FAILURE);
+	}
 	ft_bzero(mem, size * count);
 	return ((void *) mem);
 }
