+++ minishell/libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 15:42:10 by seunan            #+#    #+#             */
-/*   Updated: 2023/08/27 04:06:19 by seunan           ###   ########.fr       */
+/*   Updated: 2023/09/04 22:24:49 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	i = 0;
 	if (!s1)
 	{
-		ft_putstr_fd("exit\n", 1);
+		ft_putstr_fd("exit\n", 2);
 		exit(EXIT_FAILURE);
 	}
 	while (i < n && !(s1[i] == '\0' && s2[i] == '\0'))
