+++ minishell/libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 16:29:19 by seunan            #+#    #+#             */
-/*   Updated: 2023/08/27 03:46:54 by seunan           ###   ########.fr       */
+/*   Updated: 2023/09/04 22:18:10 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
