+++ minishell/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 11:29:31 by seunan            #+#    #+#             */
-/*   Updated: 2023/09/04 16:21:31 by seunan           ###   ########.fr       */
+/*   Updated: 2023/09/04 22:18:15 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
